#27 - moves monthly limit test to home module
